full/tests/requirements.txt
robotframework==4.1.1
robotframework-seleniumlibrary==5.1.3